Flyttede nye klasser til separate filer
ja
