+++ GirafRest.Test/Controllers/AccountControllerTest.cs
@@ -20,132 +20,13 @@ using Microsoft.Extensions.Logging;
 using ILoggerFactory = Microsoft.Extensions.Logging.ILoggerFactory;
 using System.Threading;
 using System.Linq;
+using GirafRest.Test.RepositoryMocks;
 using SignInResult = Microsoft.AspNetCore.Identity.SignInResult;
 
 namespace GirafRest.Test
 {
     public class AccountControllerTest
     {
-        public class MockUserManager : Mock<UserManager<GirafUser>>
-        {
-            public MockUserManager()
-                : this(new Mock<IUserStore<GirafUser>>())
-            { }
-
-            public MockUserManager(Mock<IUserStore<GirafUser>> userStore)
-                : this(
-                    userStore,
-                    new Mock<IOptions<IdentityOptions>>(),
-                    new Mock<IPasswordHasher<GirafUser>>(),
-                    new Mock<IEnumerable<IUserValidator<GirafUser>>>(),
-                    new Mock<IEnumerable<IPasswordValidator<GirafUser>>>(),
-                    new Mock<ILookupNormalizer>(),
-                    new Mock<IdentityErrorDescriber>(),
-                    new Mock<IServiceProvider>(),
-                    new Mock<ILogger<UserManager<GirafUser>>>()
-                )
-            { }
-
-            public MockUserManager(
-                    Mock<IUserStore<GirafUser>> userStore,
-                    Mock<IOptions<IdentityOptions>> options,
-                    Mock<IPasswordHasher<GirafUser>> passwordHasher,
-                    Mock<IEnumerable<IUserValidator<GirafUser>>> userValidators,
-                    Mock<IEnumerable<IPasswordValidator<GirafUser>>> passwordValidators,
-                    Mock<ILookupNormalizer> lookupNormalizers,
-                    Mock<IdentityErrorDescriber> identityErrorDescriber,
-                    Mock<IServiceProvider> serviceProvider,
-                    Mock<ILogger<UserManager<GirafUser>>> logger)
-                : base(
-                    userStore.Object,
-                    options.Object,
-                    passwordHasher.Object,
-                    null,
-                    null,
-                    lookupNormalizers.Object,
-                    identityErrorDescriber.Object,
-                    serviceProvider.Object,
-                    logger.Object
-                )
-            {
-                UserStore = userStore;
-                Options = options;
-                PasswordHasher = passwordHasher;
-                UserValidators = userValidators;
-                PasswordValidators = passwordValidators;
-                LookupNormalizers = lookupNormalizers;
-                IdentityErrorDescriber = identityErrorDescriber;
-                ServiceProvider = serviceProvider;
-                Logger = logger;
-            }
-
-            public Mock<IUserStore<GirafUser>> UserStore { get; }
-            public Mock<IOptions<IdentityOptions>> Options { get; }
-            public Mock<IPasswordHasher<GirafUser>> PasswordHasher { get; }
-            public Mock<IEnumerable<IUserValidator<GirafUser>>> UserValidators { get; }
-            public Mock<IEnumerable<IPasswordValidator<GirafUser>>> PasswordValidators { get; }
-            public Mock<ILookupNormalizer> LookupNormalizers { get; } 
-            public Mock<IdentityErrorDescriber> IdentityErrorDescriber { get; }
-            public Mock<IServiceProvider> ServiceProvider { get; }
-            public Mock<ILogger<UserManager<GirafUser>>> Logger { get; }
-        }
-        
-        public class MockSignInManager : Mock<SignInManager<GirafUser>>
-        {
-            public MockSignInManager()
-                : this(new MockUserManager())
-            { }
-
-            public MockSignInManager(
-                Mock<UserManager<GirafUser>> userManager
-            )
-                : this(
-                    userManager,
-                    new Mock<IHttpContextAccessor>(),
-                    new Mock<IUserClaimsPrincipalFactory<GirafUser>>(),
-                    new Mock<IOptions<IdentityOptions>>(),
-                    new Mock<ILogger<SignInManager<GirafUser>>>(),
-                    new Mock<IAuthenticationSchemeProvider>(),
-                    new Mock<IUserConfirmation<GirafUser>>())
-            { }
-
-            public MockSignInManager(
-                    Mock<UserManager<GirafUser>> userManager,
-                    Mock<IHttpContextAccessor> httpContextAccessor,
-                    Mock<IUserClaimsPrincipalFactory<GirafUser>> userClaimsPrincipalFactory,
-                    Mock<IOptions<IdentityOptions>> options,
-                    Mock<ILogger<SignInManager<GirafUser>>> logger,
-                    Mock<IAuthenticationSchemeProvider> authenticationSchemeProvider,
-                    Mock<IUserConfirmation<GirafUser>> userConfirmation
-                )
-                : base (
-                    userManager.Object,
-                    httpContextAccessor.Object,
-                    userClaimsPrincipalFactory.Object,
-                    options.Object,
-                    logger.Object,
-                    authenticationSchemeProvider.Object,
-                    userConfirmation.Object
-                )
-            {
-                UserManager = userManager;
-                IHttpContextAccessor = httpContextAccessor;
-                IUserClaimsPrincipalFactory = userClaimsPrincipalFactory;
-                IOptions = options;
-                ILogger = logger;
-                IAuthenticationSchemeProvider = authenticationSchemeProvider;
-                IUserConfirmation = userConfirmation;
-            }
-
-            public Mock<UserManager<GirafUser>> UserManager { get; }
-            public Mock<IHttpContextAccessor> IHttpContextAccessor { get; }
-            public Mock<IUserClaimsPrincipalFactory<GirafUser>> IUserClaimsPrincipalFactory { get; }
-            public Mock<IOptions<IdentityOptions>> IOptions { get; }
-            public Mock<ILogger<SignInManager<GirafUser>>> ILogger { get; }
-            public Mock<IAuthenticationSchemeProvider> IAuthenticationSchemeProvider { get; }
-            public Mock<IUserConfirmation<GirafUser>> IUserConfirmation { get; }
-        }
-
         public class OptionsJwtConfig : IOptions<JwtConfig>
         {
             public JwtConfig Value { get; }
@@ -155,86 +36,6 @@ namespace GirafRest.Test
                 Value = config;
             }
         }
-
-        public class MockedAccountController : AccountController
-        {
-            public MockedAccountController()
-                : this(new MockSignInManager())
-            { }
-
-            public MockedAccountController(IOptions<JwtConfig> configuration) 
-                : this(
-                    new MockSignInManager(), 
-                    configuration
-                )
-            { }
-
-            public MockedAccountController(MockSignInManager signInManager) 
-                : this(
-                    signInManager, 
-                    new OptionsJwtConfig(default)
-                )
-            { }
-
-            public MockedAccountController(
-                MockSignInManager signInManager,
-                IOptions<JwtConfig> configuration)
-                : this(
-                    signInManager, 
-                    new Mock<ILoggerFactory>(),
-                    new Mock<IGirafService>(),
-                    configuration,
-                    new Mock<IGirafUserRepository>(),
-                    new Mock<IDepartmentRepository>(),
-                    new Mock<IGirafRoleRepository>()
-                )
-            { }
-            
-            public MockedAccountController(
-                Mock<SignInManager<GirafUser>> signInManager,
-                Mock<ILoggerFactory> loggerFactory,
-                Mock<IGirafService> giraf,
-                IOptions<JwtConfig> configuration,
-                Mock<IGirafUserRepository> userRepository,
-                Mock<IDepartmentRepository> departmentRepository, 
-                Mock<IGirafRoleRepository> girafRoleRepository
-                ) 
-                : base(
-                    signInManager.Object, 
-                    loggerFactory.Object,
-                    giraf.Object, 
-                    configuration,
-                    userRepository.Object,
-                    departmentRepository.Object,
-                    girafRoleRepository.Object
-                )
-            {
-                SignInManager = signInManager;
-                LoggerFactory = loggerFactory;
-                GirafService = giraf;
-                UserRepository = userRepository;
-                DepartmentRepository = departmentRepository;
-                GirafRoleRepository = girafRoleRepository;
-
-                // The following are primary mocks whcih are generic.
-                //   These are added to ease the development of tests.
-                var affectedRows = 1;
-                GirafService.Setup(
-                    service => service._context.SaveChangesAsync(It.IsAny<CancellationToken>())
-                ).Returns(Task.FromResult(affectedRows));
-                GirafService.SetupGet(
-                    service => service._userManager
-                ).Returns(signInManager.Object.UserManager);
-            }
-            
-            public Mock<SignInManager<GirafUser>> SignInManager { get; }
-            public Mock<ILoggerFactory> LoggerFactory { get; }
-            public Mock<IGirafService> GirafService { get; }
-            public Mock<IGirafUserRepository> UserRepository { get; }
-            public Mock<IDepartmentRepository> DepartmentRepository { get; }
-            public Mock<IGirafRoleRepository> GirafRoleRepository { get; }
-        }
-
         
         #region Login
